feat: separate getDMChannelTitle util function and modify message editor's placeholder in thread list
